Theme.json: Remove template part data from parent theme.
Having the partials set up with area data is causing the template-part block to use that tagName when rendering the template. This causes issues for our header and footer blocks, which each use `header` and `footer` tags themselves.

This removes the other template partials as well, because they're not being used in the child themes.

See https://github.com/WordPress/five-for-the-future/issues/325.